requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Admin==1.6.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
